plotdap: Easily Visualize Data from 'ERDDAP' Servers via the 'rerddap' Package

Easily visualize and animate 'tabledap' and 'griddap' objects obtained via the 'rerddap' package in a simple one-line command, using either base graphics or 'ggplot2' graphics. 'plotdap' handles extracting and reshaping the data, map projections and continental outlines. Optionally the data can be animated through time using the 'gganmiate' package.#'

Details

The package contains four main functions:

Main Functions

plotdap - Visualize data returned from rerddap servers.

add_griddap - Add rerddap::griddap() data to a plotdap map

add_tabledap - Add rerddap::tabledap data to a plotdap map

add_ggplot - Add ggplot2 elements to a plotdap object

#' @section Details: Besides the terse help documents, more detail in using the functions are given in the included vignette "using_plotdap". The datasets used in the vignette are included in the "data" directory.

See also

Author

Maintainer: Roy Mendelssohn roy.mendelssohn@noaa.gov [contributor]

Authors:

Carson Sievert